+++ cyber/logger/async_logger.h
@@ -139,7 +139,8 @@ class AsyncLogger : public google::base::Logger {
     }
 
     inline void add(Msg&& msg, bool force_flush) {
-      size += sizeof(msg) + msg.message.size();
+      size += static_cast<int>(sizeof(msg))
+              + static_cast<int>(msg.message.size());
       messages.emplace_back(std::move(msg));
       flush |= force_flush;
     }
